Cleanup: AST: 6: Relocate location_test.rs into location.rs
Move tests from location_test.rs into location.rs using a #[cfg(test)]
module. This follows the idiomatic Rust convention of co-locating unit
tests with the code they test. Delete location_test.rs.
